+++ tests/src/main_tests.rs
@@ -2,6 +2,7 @@ use std::{sync::Arc, time::Duration};
 
 use miden_client::{
     ClientError, ONE,
+    account::AccountId,
     builder::ClientBuilder,
     keystore::FilesystemKeyStore,
     rpc::{Endpoint, NodeRpcClient, TonicRpcClient, domain::account::FetchedAccount},
@@ -1198,4 +1199,35 @@ async fn ignore_invalid_notes() {
     assert_eq!(consumed_notes.len(), 2);
     assert!(consumed_notes.iter().any(|note| note.id() == note_1.id()));
     assert!(consumed_notes.iter().any(|note| note.id() == note_2.id()));
+}
+
+#[tokio::test]
+async fn output_only_note() {
+    let (mut client, authenticator) = create_test_client().await;
+
+    let faucet =
+        insert_new_fungible_faucet(&mut client, AccountStorageMode::Private, &authenticator)
+            .await
+            .unwrap()
+            .0;
+
+    let fungible_asset = FungibleAsset::new(faucet.id(), MINT_AMOUNT).unwrap();
+    let tx_request = TransactionRequestBuilder::new()
+        .build_mint_fungible_asset(
+            fungible_asset,
+            AccountId::try_from(ACCOUNT_ID_REGULAR).unwrap(),
+            NoteType::Public,
+            client.rng(),
+        )
+        .unwrap();
+    let note_id = tx_request.expected_output_own_notes().pop().unwrap().id();
+    execute_tx_and_sync(&mut client, fungible_asset.faucet_id(), tx_request.clone()).await;
+
+    // The created note should be an output only note because it is not consumable by any client
+    // account.
+    let input_note = client.get_input_note(note_id).await.unwrap();
+    assert!(input_note.is_none());
+
+    let output_note = client.get_output_note(note_id).await.unwrap();
+    assert!(output_note.is_some());
 }